Rename lib folder to util to prevent dist files conflict
